Compiler warnings cleanup: Use size_t instead on int as loop counter in sipXtapiTestSuite::testSendInfoExternalTransport().
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9745 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXcallLib/src/test/tapi/sipXtapiTestCall.cpp
@@ -1779,7 +1779,7 @@ void sipXtapiTestSuite::testSendInfoExternalTransport()
         memset(aAddress, 0, 32);
         sipxConfigGetLocalContacts(g_hInst, aAddress, 32, nDummy);
 
-        for (int j = 0; j < nDummy; j++)
+        for (size_t j = 0; j < nDummy; j++)
         {
            if (strcmp(aAddress[j].cCustomTransportName, "info-tunnel") == 0)
            {
